Warn before disabling SSL verification, clean-up whois requests.
It seems that sometimes raw WHOIS data comes back in a PRE tag with
the id of "registryData" and other times "registrarData", try both
to increase likelihood of success.

Also, replace '\r\n' in content with '\n'. Note: This should probably
only be done on non-windows platforms but for now, I'm working on the
idea that the plugin is "only supported on Linux".